// frontend/frontend/src/swagger/model/draftFvciIncomeDetailsDto.ts
/**
 * NSDL API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface DraftFvciIncomeDetailsDto { 
    id?: number;
    fvciApplicationId?: string;
    sourceOfIncome?: Array<string>;
    codeOfBusiness?: string;
    grossAnnualIncome?: number;
    netWorth?: number;
    asOnDate?: Date;
    status?: number;
    createdAt?: Date;
    updatedAt?: Date;
}